Absolutely massive commit:
Most manually installed packages have been removed, and replaced with ELPA/MELPA package.el based packages. Hacks (where neccessary.) have been moved out to plugin, style function librar(y|ies) or left for legacy, or ditched. YaSnippet remains as the only Major library that is still manually installed (because frankly I don't have time to vet all the snippets.)

I've also added a few snippets for modes I'm using a lot for the last few months, mostly a few KendoUI Mobile helpers, and some other javascript and html stuff.

Autopair mode has been dropped in favor of Flex-autopair, it's not a complete overlap of functionality, but flex-autopair feels a little better, honestly though, after using the new XCode a lot at the Tail end of 2012, I'm still quite jealous of the paren handling in that, which I think is the best I've seen.

A huge number of additional libraries have been installed, but thanks to auto-load, startup time has been significantly improved.

Notable (?) libraries added are:

* Elnode
* Rest Client
* Slime
* Helm
* Expand Region
* Multiple Cursors
* IEdit
* FlyMake / FlyCheck
* Several inf/repl modes, (Ruby in particular, not had much time to wrangle SwankJS or some other JS repl, but running a node shell in a window isn't much different, I'd quite like to get my Chromium dev tools somehow hooked into Emacs, that would be cool, but I'm not sure it's do-able.)
* A few google tools, contact, google at point etc.
* Github gist (wasn't working the last version I had, seems ok, no, GOOD now.)
* Slim, less, coffee
* ...
* Oh, just so many more, Package.el with Melpa is getting much better now, mostly things load/install successfully, which wasn't how it was before.

Anyway, a ton of stuff, so if you are actually reading this, remember this is primarily just for me, but if you want to try it out, backup your config, and just clone this into `.emacs.d` and play.

This is officially the "My God! It's full of stars!" config.