First:
An apology to John Olsson
=========================

for inadvertantly re-badging Subatomic-atomic and placing it online as my own work, this hadn't meant to end up online, and was part of a big commit of my .Emacs.d - I've fixed the file header to include full credit for John and offer my apologies for the error.

Now, a couple of patches to my init.el

Flymake Modals?
===============

Flymake, turn off those modal dialogs for goodness sake, I want to type here, not click stupid buttons!  So you borked, and you want to tell me about it? Cute, but seriously DON'T default to a modal, there's nothing more wretched. Ok, so that's a massive overreaction, but without honest feedback, who knows what the truth is?

It's good that we can change it, but still, bleurgh!

Anyway, changed in the init.el.

TODO, FIXME etc..
=================

Added a quick hack to the `prog-modes-hook` which will highlight `TODO:`, `FIXME:`, `HACK:`, `NOTE:`, `BUG:`, `REFACTOR:` and `THE HORROR:` to the `font-lock-warning-face`.

Obviously `THE HORROR:` will be the one to use on my very worst code offenses, and I hope that catches on. All fledgling coders who've just discovered tail-recursion and vow to use it for ALL their loops, take note.