Update Packages; Disable Company Mode; Update Pivotal init
Aside from the regular package update, this commit also contains
a revert from Company Mode.

I gave company mode a try a while back.  Whether or not I made any comment
about it at that time or not, I'm not sure.  To recap I was less than
impressed,  I found AC to be fundamentally better than Company mode, in
all things.

There seems to be a few overly zealous supporters of Company mode, my
initial impression is, well good for you. It's nice to enjoy the tools
you use.

My second impression is that, maybe, just maybe, you never really gave
AutoComplete a chance, or perhaps you were more concerned with things
that aren't important to me.  I expect the inverse is also true.

For me things like the popup render in AC (which isn't all that
marvellous) is still more than adequate, when it does screw up, it's
usually because I have text magnification on or somethings.  Admittedly
I don't see why that would happend, but I'm not overly concerned.

The only other thing that troubled me about AC is that it sometimes
stepped on YASnippet's toes, and I couldn't get a TAB expansion when I
needed it.

I can solve this by stopping AC from activating for a few extra ms, so
that my usual flow of YASnippet is uniterrupted.

Anyway, I feel this note is already too long, so let's not get
side-tracked.

My issues with Company mode, on the other hand are to do with core
functionality, something AC has never let me down on.

For example, if I'm working in Emacs Lisp mode, (which should have the
best support in Emacs, right?)   I expect to see a popup menu the moment
I hesitate on a command name...

... So I wait, and nothing.

Same for many modes, anyway I know the thing is on, because I see it
from time to time, at odd and unpredicatble moments.

Which is _obviously_ very handy.

Anywhoo that's not all that inspiring, and I don't want to waste time
with it.  If you think I set things up wrong, please let me know, I have
the relavant setup details in ~/.emacs.d/modes-init/init-company.el.

What this has led me to do is delve deeper into the configuration
possibilities of both AC and Company mode.

Let me just say, AC is better.  YMMV of course, but consider this the
thoughts held after several years considering the merits of both.  It's
up to you to decide if that's worth anything.

Pivotal
------------

Ok, so I also added better support for auto-discovery of a Pivotal API
key in the current system environment or in a ~/.pivotal_api_key file.

Either way, if it's there it will be found in startup, and you can
access your pivotal projects with M-x pivotal-get-projects